use expected score to select which move to play
Instead of just playing the move that does best
against the predicted one, use all the information available
in the predicted probabilities to select the one that
optimizes the expected score. E.g. when the predicted
probabilities for opponent move are (0.1,0.44,0.46) play
scissors instead of rock.
Then we use these scores as logits and apply temperature
as before (so moves are not deterministic).